fix assembly or frequency
