+++ lib/constants.js
@@ -1,215 +1,1302 @@
-export const DATASETS = {
-  "gem_j_wga": {
-    "id": "gem_j_wga",
-    "idx": 1,
-    "source": "GEM-J WGA",
+export const DATASETS = [
+  {
     "dataset": () => "GEM-J WGA",
-    "population": "Japanese",
+    "layer1": "Japanese",
+    "layer2": "",
+    "layer3": "",
+    "source": "gem_j_wga"
   },
-  "jga_ngs": {
-    "id": "jga_ngs",
-    "idx": 2,
-    "source": "JGA-NGS",
+  {
     "dataset": () => "JGA-NGS",
-    "population": "Japanese",
+    "layer1": "Japanese",
+    "layer2": "",
+    "layer3": "",
+    "source": "jga_ngs"
   },
-  "jga_snp": {
-    "id": "jga_snp",
-    "idx": 3,
-    "source": "JGA-SNP",
+  {
     "dataset": () => "JGA-SNP",
-    "population": "Japanese",
+    "layer1": "Total",
+    "layer2": "",
+    "layer3": "",
+    "source": "jga_snp"
   },
-  "tommo": {
-    "id": "tommo",
-    "idx": 4,
-    "source": "ToMMo",
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Lung cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno1"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Lung cancer",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno1.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Lung cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno1.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Esophageal cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno2"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Esophageal cancer",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno2.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Esophageal cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno2.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Gastric cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno3"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Gastric cancer",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno3.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Gastric cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno3.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Colorectal cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno4"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Colorectal cancer",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno4.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Colorectal cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno4.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Liver cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno5"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Liver cancer",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno5.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Liver cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno5.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Pancreas cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno6"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Pancreas cancer",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno6.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Pancreas cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno6.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Gallbladder/Cholangiocarcinoma",
+    "layer3": "",
+    "source": "bbj_riken.mpheno7"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Gallbladder/Cholangiocarcinoma",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno7.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Gallbladder/Cholangiocarcinoma",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno7.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Prostate cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno8"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Prostate cancer",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno8.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Breast cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno9"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Breast cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno9.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Cervical cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno10"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Cervical cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno10.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Uterine cancer",
+    "layer3": "",
+    "source": "bbj_riken.mpheno11"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Uterine cancer",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno11.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Ovarian center",
+    "layer3": "",
+    "source": "bbj_riken.mpheno12"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Ovarian center",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno12.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Hematopoietic tumor",
+    "layer3": "",
+    "source": "bbj_riken.mpheno13"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Hematopoietic tumor",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno13.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Neoplasms",
+    "layer2": "Hematopoietic tumor",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno13.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Cerebral infarction",
+    "layer3": "",
+    "source": "bbj_riken.mpheno14"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Cerebral infarction",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno14.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Cerebral infarction",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno14.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Cerebral aneurysm",
+    "layer3": "",
+    "source": "bbj_riken.mpheno15"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Cerebral aneurysm",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno15.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Cerebral aneurysm",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno15.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Epilepsy",
+    "layer3": "",
+    "source": "bbj_riken.mpheno16"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Epilepsy",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno16.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the nervous system",
+    "layer2": "Epilepsy",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno16.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Bronchial asthma",
+    "layer3": "",
+    "source": "bbj_riken.mpheno17"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Bronchial asthma",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno17.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Bronchial asthma",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno17.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Pulmonary tuberculosis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno18"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Pulmonary tuberculosis",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno18.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Pulmonary tuberculosis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno18.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Chronic obstructive pulmonary disease",
+    "layer3": "",
+    "source": "bbj_riken.mpheno19"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Chronic obstructive pulmonary disease",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno19.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Chronic obstructive pulmonary disease",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno19.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Interstitial lung disease/Pulmonary fibrosis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno20"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Interstitial lung disease/Pulmonary fibrosis",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno20.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the respiratory system",
+    "layer2": "Interstitial lung disease/Pulmonary fibrosis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno20.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Myocardial infarction",
+    "layer3": "",
+    "source": "bbj_riken.mpheno21"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Myocardial infarction",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno21.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Myocardial infarction",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno21.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Unstable angina",
+    "layer3": "",
+    "source": "bbj_riken.mpheno22"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Unstable angina",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno22.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Unstable angina",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno22.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Stable angina",
+    "layer3": "",
+    "source": "bbj_riken.mpheno23"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Stable angina",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno23.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Stable angina",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno23.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Arrhythmia",
+    "layer3": "",
+    "source": "bbj_riken.mpheno24"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Arrhythmia",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno24.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Arrhythmia",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno24.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Heart failure",
+    "layer3": "",
+    "source": "bbj_riken.mpheno25"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Heart failure",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno25.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Heart failure",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno25.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Peripheral arterial diseases",
+    "layer3": "",
+    "source": "bbj_riken.mpheno26"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Peripheral arterial diseases",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno26.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the circulatory system",
+    "layer2": "Peripheral arterial diseases",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno26.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Chronic hepatitis B",
+    "layer3": "",
+    "source": "bbj_riken.mpheno27"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Chronic hepatitis B",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno27.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Chronic hepatitis B",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno27.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Chronic hepatitis C",
+    "layer3": "",
+    "source": "bbj_riken.mpheno28"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Chronic hepatitis C",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno28.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Chronic hepatitis C",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno28.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Liver cirrhosis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno29"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Liver cirrhosis",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno29.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Hepato-Biliary-Pancreatic diseases",
+    "layer2": "Liver cirrhosis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno29.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the genitourinary system",
+    "layer2": "Nephrotic syndrome",
+    "layer3": "",
+    "source": "bbj_riken.mpheno30"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the genitourinary system",
+    "layer2": "Nephrotic syndrome",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno30.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the genitourinary system",
+    "layer2": "Nephrotic syndrome",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno30.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the genitourinary system",
+    "layer2": "Urolithiasis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno31"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the genitourinary system",
+    "layer2": "Urolithiasis",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno31.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the genitourinary system",
+    "layer2": "Urolithiasis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno31.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Osteoporosis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno32"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Osteoporosis",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno32.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Osteoporosis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno32.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Diabetes mellitus",
+    "layer3": "",
+    "source": "bbj_riken.mpheno33"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Diabetes mellitus",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno33.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Diabetes mellitus",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno33.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Dyslipidemia",
+    "layer3": "",
+    "source": "bbj_riken.mpheno34"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Dyslipidemia",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno34.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Metabolic diseases",
+    "layer2": "Dyslipidemia",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno34.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Endocrine diseases",
+    "layer2": "Graves' disease",
+    "layer3": "",
+    "source": "bbj_riken.mpheno35"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Endocrine diseases",
+    "layer2": "Graves' disease",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno35.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Endocrine diseases",
+    "layer2": "Graves' disease",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno35.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Collagen diseases (connective tissue diseases)",
+    "layer2": "Rheumatoid arthritis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno36"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Collagen diseases (connective tissue diseases)",
+    "layer2": "Rheumatoid arthritis",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno36.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Collagen diseases (connective tissue diseases)",
+    "layer2": "Rheumatoid arthritis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno36.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Allergy",
+    "layer2": "Hay fever",
+    "layer3": "",
+    "source": "bbj_riken.mpheno37"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Allergy",
+    "layer2": "Hay fever",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno37.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Allergy",
+    "layer2": "Hay fever",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno37.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Drug eruption",
+    "layer3": "",
+    "source": "bbj_riken.mpheno38"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Drug eruption",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno38.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Drug eruption",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno38.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Atopic dermatitis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno39"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Atopic dermatitis",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno39.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Atopic dermatitis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno39.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Keloid",
+    "layer3": "",
+    "source": "bbj_riken.mpheno40"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Keloid",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno40.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the skin and subcutaneous tissue",
+    "layer2": "Keloid",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno40.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Gynecological diseases",
+    "layer2": "Uterine fibroid",
+    "layer3": "",
+    "source": "bbj_riken.mpheno41"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Gynecological diseases",
+    "layer2": "Uterine fibroid",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno41.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Gynecological diseases",
+    "layer2": "Endometriosis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno42"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Gynecological diseases",
+    "layer2": "Endometriosis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno42.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the eye and adnexa",
+    "layer2": "Glaucoma",
+    "layer3": "",
+    "source": "bbj_riken.mpheno44"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the eye and adnexa",
+    "layer2": "Glaucoma",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno44.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the eye and adnexa",
+    "layer2": "Glaucoma",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno44.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the eye and adnexa",
+    "layer2": "Cataract",
+    "layer3": "",
+    "source": "bbj_riken.mpheno45"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the eye and adnexa",
+    "layer2": "Cataract",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno45.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Diseases of the eye and adnexa",
+    "layer2": "Cataract",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno45.female"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Oral and Maxillofacial diseases",
+    "layer2": "Periodontitis",
+    "layer3": "",
+    "source": "bbj_riken.mpheno46"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Oral and Maxillofacial diseases",
+    "layer2": "Periodontitis",
+    "layer3": "Male",
+    "source": "bbj_riken.mpheno46.male"
+  },
+  {
+    "dataset": () => "JGA-SNP",
+    "layer1": "Oral and Maxillofacial diseases",
+    "layer2": "Periodontitis",
+    "layer3": "Female",
+    "source": "bbj_riken.mpheno46.female"
+  },
+  {
     "dataset": (assembly) => {
       switch (assembly) {
         case "GRCh37":
-          return "ToMMo 8.3KJPN"
+          return "ToMMo 8.3KJPN";
         case "GRCh38":
-          return "ToMMo 54KJPN"
+          return "ToMMo 54KJPN";
         default:
-          return "ToMMo"
+          return "ToMMo";
       }
     },
-    "population": "Japanese",
+    "layer1": "Japanese",
+    "layer2": "",
+    "layer3": "",
+    "source": "tommo"
   },
-  "hgvd": {
-    "id": "hgvd",
-    "idx": 5,
-    "source": "HGVD",
+  {
     "dataset": () => "HGVD",
-    "population": "Japanese",
+    "layer1": "Japanese",
+    "layer2": "",
+    "layer3": "",
+    "source": "hgvd"
   },
-  "gnomad_genomes": {
-    "id": "gnomad_genomes",
-    "idx": 6,
-    "source": "gnomAD Genomes",
+  { "dataset": () => "NCBN", "layer1": "Total", "layer2": "", "layer3": "", "source": "ncbn" },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Japanese",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.jpn"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Japanese",
+    "layer2": "Hondo",
+    "layer3": "",
+    "source": "ncbn.jpn.hondo"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Japanese",
+    "layer2": "Ryukyu",
+    "layer3": "",
+    "source": "ncbn.jpn.ryukyu"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "African Caribbean in Barbados",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.acb"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "African Ancestry in SW USA",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.asw"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Bengali in Bangladesh",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.beb"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "British From England and Scotland",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.gbr"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Chinese Dai in Xishuangbanna, China",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.cdx"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Northern and Western European",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.ceu"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Colombian in Medellín, Colombia",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.clm"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Esan in Nigeria",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.esn"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Finnish in Finland",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.fin"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Gambian in Western Division – Mandinka",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.gwd"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Gujarati Indians in Houston, Texas, USA",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.gih"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Han Chinese in Beijing, China",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.chb"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Han Chinese South",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.chs"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Iberian Populations in Spain",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.ibs"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Indian Telugu in the U.K.",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.itu"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Japanese in Tokyo, Japan",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.jpt"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Kinh in Ho Chi Minh City, Vietnam",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.khv"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Luhya in Webuye, Kenya",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.lwk"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Mende in Sierra Leone",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.msl"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Mexican Ancestry in Los Angeles CA USA",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.mxl"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Peruvian in Lima Peru",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.pel"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Puerto Rican in Puerto Rico",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.pur"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Punjabi in Lahore, Pakistan",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.pjl"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Sri Lankan Tamil in the UK",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.stu"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Toscani in Italia",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.tsi"
+  },
+  {
+    "dataset": () => "NCBN",
+    "layer1": "Yoruba in Ibadan, Nigeria",
+    "layer2": "",
+    "layer3": "",
+    "source": "ncbn.yri"
+  },
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "Total",
+    "layer1": "Total",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes"
   },
-  "gnomad_genomes_african": {
-    "id": "gnomad_genomes_african",
-    "idx": 7,
-    "source": "gnomAD Genomes:African-American/African",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "African-American/African",
+    "layer1": "African-American/African",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.afr"
   },
-  "gnomad_genomes_amish": {
-    "id": "gnomad_genomes_amish",
-    "idx": 8,
-    "source": "gnomAD Genomes:Amish",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "Amish",
+    "layer1": "Amish",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.ami"
   },
-  "gnomad_genomes_latino": {
-    "id": "gnomad_genomes_latino",
-    "idx": 9,
-    "source": "gnomAD Genomes:Latino",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "Latino",
+    "layer1": "Latino",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.amr"
   },
-  "gnomad_genomes_ashkenazi_jewish": {
-    "id": "gnomad_genomes_ashkenazi_jewish",
-    "idx": 10,
-    "source": "gnomAD Genomes:Ashkenazi Jewish",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "Ashkenazi Jewish",
+    "layer1": "Ashkenazi Jewish",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.asj"
   },
-  "gnomad_genomes_east_asian": {
-    "id": "gnomad_genomes_east_asian",
-    "idx": 11,
-    "source": "gnomAD Genomes:East Asian",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "East Asian",
+    "layer1": "East Asian",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.eas"
   },
-  "gnomad_genomes_european_finnish": {
-    "id": "gnomad_genomes_european_finnish",
-    "idx": 12,
-    "source": "gnomAD Genomes:Finnish",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "European (Finnish)",
+    "layer1": "European (Finnish)",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.fin"
   },
-  "gnomad_genomes_middle_eastern": {
-    "id": "gnomad_genomes_middle_eastern",
-    "idx": 13,
-    "source": "gnomAD Genomes:Middle Eastern",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "Middle Eastern",
+    "layer1": "Middle Eastern",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.mid"
   },
-  "gnomad_genomes_european_non_finnish": {
-    "id": "gnomad_genomes_european_non_finnish",
-    "idx": 14,
-    "source": "gnomAD Genomes:Non-Finnish European",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "European (Non-Finnish)",
+    "layer1": "European (Non-Finnish)",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.nfe"
   },
-  "gnomad_genomes_other": {
-    "id": "gnomad_genomes_other",
-    "idx": 16,
-    "source": "gnomAD Genomes:Other",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "Other",
+    "layer1": "Remaining",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.remaining"
   },
-  "gnomad_genomes_remaining": {
-    "id": "gnomad_genomes_remaining",
-    "idx": 17,
-    "source": "gnomAD Genomes:Remaining",
+  {
     "dataset": () => "gnomAD Genomes",
-    "population": "Remaining",
+    "layer1": "South Asian",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_genomes.sas"
   },
-  "gnomad_genomes_south_asian": {
-    "id": "gnomad_genomes_south_asian",
-    "idx": 18,
-    "source": "gnomAD Genomes:South Asian",
-    "dataset": () => "gnomAD Genomes",
-    "population": "South Asian",
-  },
-  "gnomad_exomes": {
-    "id": "gnomad_exomes",
-    "idx": 19,
-    "source": "gnomAD Exomes",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "Total",
+    "layer1": "Total",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes"
   },
-  "gnomad_exomes_african": {
-    "id": "gnomad_exomes_african",
-    "idx": 20,
-    "source": "gnomAD Exomes:African-American/African",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "African-American/African",
+    "layer1": "African-American/African",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.afr"
   },
-  "gnomad_exomes_latino": {
-    "id": "gnomad_exomes_latino",
-    "idx": 21,
-    "source": "gnomAD Exomes:Latino",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "Latino",
+    "layer1": "Latino",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.amr"
   },
-  "gnomad_exomes_ashkenazi_jewish": {
-    "id": "gnomad_exomes_ashkenazi_jewish",
-    "idx": 22,
-    "source": "gnomAD Exomes:Ashkenazi Jewish",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "Ashkenazi Jewish",
+    "layer1": "Ashkenazi Jewish",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.asj"
   },
-  "gnomad_exomes_east_asian": {
-    "id": "gnomad_exomes_east_asian",
-    "idx": 23,
-    "source": "gnomAD Exomes:East Asian",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "East Asian",
+    "layer1": "East Asian",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.eas"
   },
-  "gnomad_exomes_european_finnish": {
-    "id": "gnomad_exomes_european_finnish",
-    "idx": 24,
-    "source": "gnomAD Exomes:Finnish",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "European (Finnish)",
+    "layer1": "European (Finnish)",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.fin"
   },
-  "gnomad_exomes_middle_eastern": {
-    "id": "gnomad_exomes_middle_eastern",
-    "idx": 25,
-    "source": "gnomAD Exomes:Middle Eastern",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "Middle Eastern",
+    "layer1": "Middle Eastern",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.mid"
   },
-  "gnomad_exomes_european_non_finnish": {
-    "id": "gnomad_exomes_european_non_finnish",
-    "idx": 26,
-    "source": "gnomAD Exomes:Non-Finnish European",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "European (Non-Finnish)",
+    "layer1": "European (Non-Finnish)",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.nfe"
   },
-  "gnomad_exomes_other": {
-    "id": "gnomad_exomes_other",
-    "idx": 27,
-    "source": "gnomAD Exomes:Other",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "Other",
+    "layer1": "Remaining",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.remaining"
   },
-  "gnomad_exomes_remaining": {
-    "id": "gnomad_exomes_remaining",
-    "idx": 28,
-    "source": "gnomAD Exomes:Remaining",
+  {
     "dataset": () => "gnomAD Exomes",
-    "population": "Remaining",
-  },
-  "gnomad_exomes_south_asian": {
-    "id": "gnomad_exomes_south_asian",
-    "idx": 29,
-    "source": "gnomAD Exomes:South Asian",
-    "dataset": () => "gnomAD Exomes",
-    "population": "South Asian",
-  },
-};
+    "layer1": "South Asian",
+    "layer2": "",
+    "layer3": "",
+    "source": "gnomad_exomes.sas"
+  }
+]
+
 
 export const ensureAllDatasets = frequencies => {
   ["gem_j_wga", "jga_ngs", "jga_snp", "tommo", "hgvd", "gnomad_genomes", "gnomad_exomes"].forEach(source => {
     if (!frequencies.find(x => x.source === source)) {
-      frequencies.push({source: source});
+      frequencies.push({ source: source });
     }
   });
