--- lib/constantsTest.js
@@ -1,1690 +1,0 @@
-export const DATASETS = [
-  {
-    "dataset": "GEM-J WGA",
-    "layer1": "Japanese",
-    "layer2": "",
-    "layer3": "",
-    "source": "gem_j_wga"
-  },
-  {
-    "dataset": "JGA-NGS",
-    "layer1": "Japanese",
-    "layer2": "",
-    "layer3": "",
-    "source": "jga_ngs"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Total",
-    "layer2": "",
-    "layer3": "",
-    "source": "jga_snp"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Lung cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno1"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Lung cancer",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno1.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Lung cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno1.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Esophageal cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno2"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Esophageal cancer",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno2.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Esophageal cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno2.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Gastric cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno3"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Gastric cancer",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno3.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Gastric cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno3.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Colorectal cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno4"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Colorectal cancer",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno4.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Colorectal cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno4.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Liver cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno5"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Liver cancer",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno5.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Liver cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno5.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Pancreas cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno6"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Pancreas cancer",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno6.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Pancreas cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno6.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Gallbladder/Cholangiocarcinoma",
-    "layer3": "",
-    "source": "bbj_riken.mpheno7"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Gallbladder/Cholangiocarcinoma",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno7.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Gallbladder/Cholangiocarcinoma",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno7.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Prostate cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno8"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Prostate cancer",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno8.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Breast cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno9"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Breast cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno9.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Cervical cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno10"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Cervical cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno10.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Uterine cancer",
-    "layer3": "",
-    "source": "bbj_riken.mpheno11"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Uterine cancer",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno11.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Ovarian center",
-    "layer3": "",
-    "source": "bbj_riken.mpheno12"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Ovarian center",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno12.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Hematopoietic tumor",
-    "layer3": "",
-    "source": "bbj_riken.mpheno13"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Hematopoietic tumor",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno13.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Neoplasms",
-    "layer2": "Hematopoietic tumor",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno13.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Cerebral infarction",
-    "layer3": "",
-    "source": "bbj_riken.mpheno14"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Cerebral infarction",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno14.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Cerebral infarction",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno14.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Cerebral aneurysm",
-    "layer3": "",
-    "source": "bbj_riken.mpheno15"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Cerebral aneurysm",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno15.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Cerebral aneurysm",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno15.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Epilepsy",
-    "layer3": "",
-    "source": "bbj_riken.mpheno16"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Epilepsy",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno16.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the nervous system",
-    "layer2": "Epilepsy",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno16.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Bronchial asthma",
-    "layer3": "",
-    "source": "bbj_riken.mpheno17"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Bronchial asthma",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno17.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Bronchial asthma",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno17.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Pulmonary tuberculosis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno18"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Pulmonary tuberculosis",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno18.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Pulmonary tuberculosis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno18.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Chronic obstructive pulmonary disease",
-    "layer3": "",
-    "source": "bbj_riken.mpheno19"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Chronic obstructive pulmonary disease",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno19.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Chronic obstructive pulmonary disease",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno19.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Interstitial lung disease/Pulmonary fibrosis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno20"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Interstitial lung disease/Pulmonary fibrosis",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno20.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the respiratory system",
-    "layer2": "Interstitial lung disease/Pulmonary fibrosis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno20.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Myocardial infarction",
-    "layer3": "",
-    "source": "bbj_riken.mpheno21"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Myocardial infarction",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno21.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Myocardial infarction",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno21.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Unstable angina",
-    "layer3": "",
-    "source": "bbj_riken.mpheno22"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Unstable angina",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno22.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Unstable angina",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno22.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Stable angina",
-    "layer3": "",
-    "source": "bbj_riken.mpheno23"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Stable angina",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno23.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Stable angina",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno23.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Arrhythmia",
-    "layer3": "",
-    "source": "bbj_riken.mpheno24"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Arrhythmia",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno24.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Arrhythmia",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno24.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Heart failure",
-    "layer3": "",
-    "source": "bbj_riken.mpheno25"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Heart failure",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno25.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Heart failure",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno25.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Peripheral arterial diseases",
-    "layer3": "",
-    "source": "bbj_riken.mpheno26"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Peripheral arterial diseases",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno26.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the circulatory system",
-    "layer2": "Peripheral arterial diseases",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno26.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Chronic hepatitis B",
-    "layer3": "",
-    "source": "bbj_riken.mpheno27"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Chronic hepatitis B",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno27.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Chronic hepatitis B",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno27.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Chronic hepatitis C",
-    "layer3": "",
-    "source": "bbj_riken.mpheno28"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Chronic hepatitis C",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno28.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Chronic hepatitis C",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno28.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Liver cirrhosis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno29"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Liver cirrhosis",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno29.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Hepato-Biliary-Pancreatic diseases",
-    "layer2": "Liver cirrhosis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno29.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the genitourinary system",
-    "layer2": "Nephrotic syndrome",
-    "layer3": "",
-    "source": "bbj_riken.mpheno30"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the genitourinary system",
-    "layer2": "Nephrotic syndrome",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno30.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the genitourinary system",
-    "layer2": "Nephrotic syndrome",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno30.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the genitourinary system",
-    "layer2": "Urolithiasis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno31"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the genitourinary system",
-    "layer2": "Urolithiasis",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno31.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the genitourinary system",
-    "layer2": "Urolithiasis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno31.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Osteoporosis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno32"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Osteoporosis",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno32.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Osteoporosis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno32.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Diabetes mellitus",
-    "layer3": "",
-    "source": "bbj_riken.mpheno33"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Diabetes mellitus",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno33.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Diabetes mellitus",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno33.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Dyslipidemia",
-    "layer3": "",
-    "source": "bbj_riken.mpheno34"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Dyslipidemia",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno34.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Metabolic diseases",
-    "layer2": "Dyslipidemia",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno34.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Endocrine diseases",
-    "layer2": "Graves' disease",
-    "layer3": "",
-    "source": "bbj_riken.mpheno35"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Endocrine diseases",
-    "layer2": "Graves' disease",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno35.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Endocrine diseases",
-    "layer2": "Graves' disease",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno35.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Collagen diseases (connective tissue diseases)",
-    "layer2": "Rheumatoid arthritis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno36"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Collagen diseases (connective tissue diseases)",
-    "layer2": "Rheumatoid arthritis",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno36.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Collagen diseases (connective tissue diseases)",
-    "layer2": "Rheumatoid arthritis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno36.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Allergy",
-    "layer2": "Hay fever",
-    "layer3": "",
-    "source": "bbj_riken.mpheno37"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Allergy",
-    "layer2": "Hay fever",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno37.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Allergy",
-    "layer2": "Hay fever",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno37.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Drug eruption",
-    "layer3": "",
-    "source": "bbj_riken.mpheno38"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Drug eruption",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno38.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Drug eruption",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno38.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Atopic dermatitis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno39"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Atopic dermatitis",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno39.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Atopic dermatitis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno39.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Keloid",
-    "layer3": "",
-    "source": "bbj_riken.mpheno40"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Keloid",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno40.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the skin and subcutaneous tissue",
-    "layer2": "Keloid",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno40.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Gynecological diseases",
-    "layer2": "Uterine fibroid",
-    "layer3": "",
-    "source": "bbj_riken.mpheno41"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Gynecological diseases",
-    "layer2": "Uterine fibroid",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno41.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Gynecological diseases",
-    "layer2": "Endometriosis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno42"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Gynecological diseases",
-    "layer2": "Endometriosis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno42.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the eye and adnexa",
-    "layer2": "Glaucoma",
-    "layer3": "",
-    "source": "bbj_riken.mpheno44"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the eye and adnexa",
-    "layer2": "Glaucoma",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno44.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the eye and adnexa",
-    "layer2": "Glaucoma",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno44.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the eye and adnexa",
-    "layer2": "Cataract",
-    "layer3": "",
-    "source": "bbj_riken.mpheno45"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the eye and adnexa",
-    "layer2": "Cataract",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno45.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Diseases of the eye and adnexa",
-    "layer2": "Cataract",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno45.female"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Oral and Maxillofacial diseases",
-    "layer2": "Periodontitis",
-    "layer3": "",
-    "source": "bbj_riken.mpheno46"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Oral and Maxillofacial diseases",
-    "layer2": "Periodontitis",
-    "layer3": "Male",
-    "source": "bbj_riken.mpheno46.male"
-  },
-  {
-    "dataset": "JGA-SNP",
-    "layer1": "Oral and Maxillofacial diseases",
-    "layer2": "Periodontitis",
-    "layer3": "Female",
-    "source": "bbj_riken.mpheno46.female"
-  },
-  {
-    "dataset": "ToMMo",
-    "layer1": "Japanese",
-    "layer2": "",
-    "layer3": "",
-    "source": "tommo"
-  },
-  {
-    "dataset": "HGVD",
-    "layer1": "Japanese",
-    "layer2": "",
-    "layer3": "",
-    "source": "hgvd"
-  },
-  { "dataset": "NCBN", "layer1": "Total", "layer2": "", "layer3": "", "source": "ncbn" },
-  {
-    "dataset": "NCBN",
-    "layer1": "Japanese",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.jpn"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Japanese",
-    "layer2": "Hondo",
-    "layer3": "",
-    "source": "ncbn.jpn.hondo"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Japanese",
-    "layer2": "Ryukyu",
-    "layer3": "",
-    "source": "ncbn.jpn.ryukyu"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "African Caribbean in Barbados",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.acb"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "African Ancestry in SW USA",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.asw"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Bengali in Bangladesh",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.beb"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "British From England and Scotland",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.gbr"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Chinese Dai in Xishuangbanna, China",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.cdx"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Northern and Western European",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.ceu"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Colombian in Medellín, Colombia",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.clm"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Esan in Nigeria",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.esn"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Finnish in Finland",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.fin"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Gambian in Western Division – Mandinka",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.gwd"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Gujarati Indians in Houston, Texas, USA",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.gih"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Han Chinese in Beijing, China",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.chb"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Han Chinese South",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.chs"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Iberian Populations in Spain",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.ibs"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Indian Telugu in the U.K.",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.itu"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Japanese in Tokyo, Japan",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.jpt"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Kinh in Ho Chi Minh City, Vietnam",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.khv"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Luhya in Webuye, Kenya",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.lwk"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Mende in Sierra Leone",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.msl"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Mexican Ancestry in Los Angeles CA USA",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.mxl"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Peruvian in Lima Peru",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.pel"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Puerto Rican in Puerto Rico",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.pur"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Punjabi in Lahore, Pakistan",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.pjl"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Sri Lankan Tamil in the UK",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.stu"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Toscani in Italia",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.tsi"
-  },
-  {
-    "dataset": "NCBN",
-    "layer1": "Yoruba in Ibadan, Nigeria",
-    "layer2": "",
-    "layer3": "",
-    "source": "ncbn.yri"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "Total",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "African-American/African",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.afr"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "Amish",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.ami"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "Latino",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.amr"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "Ashkenazi Jewish",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.asj"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "East Asian",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.eas"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "European (Finnish)",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.fin"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "Middle Eastern",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.mid"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "European (Non-Finnish)",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.nfe"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "Remaining",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.remaining"
-  },
-  {
-    "dataset": "gnomAD Genomes",
-    "layer1": "South Asian",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_genomes.sas"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "Total",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "African-American/African",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.afr"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "Latino",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.amr"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "Ashkenazi Jewish",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.asj"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "East Asian",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.eas"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "European (Finnish)",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.fin"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "Middle Eastern",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.mid"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "European (Non-Finnish)",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.nfe"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "Remaining",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.remaining"
-  },
-  {
-    "dataset": "gnomAD Exomes",
-    "layer1": "South Asian",
-    "layer2": "",
-    "layer3": "",
-    "source": "gnomad_exomes.sas"
-  }
-]
-
-
-export const ensureAllDatasets = frequencies => {
-  ["gem_j_wga", "jga_ngs", "jga_snp", "tommo", "hgvd", "gnomad_genomes", "gnomad_exomes"].forEach(source => {
-    if (!frequencies.find(x => x.source === source)) {
-      frequencies.push({ source: source });
-    }
-  });
-
-  return frequencies;
-};
-
-export const SO = {
-  "SO_0001483": {
-    "id": "SO_0001483",
-    "label": "SNV",
-  },
-  "SO_0000667": {
-    "id": "SO_0000667",
-    "label": "insertion",
-  },
-  "SO_0000159": {
-    "id": "SO_0000159",
-    "label": "deletion",
-  },
-  "SO_1000032": {
-    "id": "SO_1000032",
-    "label": "indel",
-  },
-  "SO_1000002": {
-    "id": "SO_1000002",
-    "label": "substitution",
-  },
-};
-
-export const CONSEQUENCE = {
-  "SO_0001893": {
-    "id": "SO_0001893",
-    "key": "transcript_ablation",
-    "label": "Transcript ablation",
-    "impact": "HIGH",
-    "order": 1,
-  },
-  "SO_0001574": {
-    "id": "SO_0001574",
-    "key": "splice_acceptor_variant",
-    "label": "Splice acceptor variant",
-    "impact": "HIGH",
-    "order": 2,
-  },
-  "SO_0001575": {
-    "id": "SO_0001575",
-    "key": "splice_donor_variant",
-    "label": "Splice donor variant",
-    "impact": "HIGH",
-    "order": 3,
-  },
-  "SO_0001587": {
-    "id": "SO_0001587",
-    "key": "stop_gained",
-    "label": "Stop gained",
-    "impact": "HIGH",
-    "order": 4,
-  },
-  "SO_0001589": {
-    "id": "SO_0001589",
-    "key": "frameshift_variant",
-    "label": "Frameshift variant",
-    "impact": "HIGH",
-    "order": 5,
-  },
-  "SO_0001578": {
-    "id": "SO_0001578",
-    "key": "stop_lost",
-    "label": "Stop lost",
-    "impact": "HIGH",
-    "order": 6,
-  },
-  "SO_0002012": {
-    "id": "SO_0002012",
-    "key": "start_lost",
-    "label": "Start lost",
-    "impact": "HIGH",
-    "order": 7,
-  },
-  "SO_0001889": {
-    "id": "SO_0001889",
-    "key": "transcript_amplification",
-    "label": "Transcript amplification",
-    "impact": "HIGH",
-    "order": 8,
-  },
-  "SO_0001821": {
-    "id": "SO_0001821",
-    "key": "inframe_insertion",
-    "label": "Inframe insertion",
-    "impact": "MODERATE",
-    "order": 9,
-  },
-  "SO_0001822": {
-    "id": "SO_0001822",
-    "key": "inframe_deletion",
-    "label": "Inframe deletion",
-    "impact": "MODERATE",
-    "order": 10,
-  },
-  "SO_0001583": {
-    "id": "SO_0001583",
-    "key": "missense_variant",
-    "label": "Missense variant",
-    "impact": "MODERATE",
-    "order": 11,
-  },
-  "SO_0001818": {
-    "id": "SO_0001818",
-    "key": "protein_altering_variant",
-    "label": "Protein altering variant",
-    "impact": "MODERATE",
-    "order": 12,
-  },
-  "SO_0001630": {
-    "id": "SO_0001630",
-    "key": "splice_region_variant",
-    "label": "Splice region variant",
-    "impact": "LOW",
-    "order": 13,
-  },
-  "SO_0001626": {
-    "id": "SO_0001626",
-    "key": "incomplete_terminal_codon_variant",
-    "label": "Incomplete terminal codon variant",
-    "impact": "LOW",
-    "order": 14,
-  },
-  "SO_0002019": {
-    "id": "SO_0002019",
-    "key": "start_retained_variant",
-    "label": "Start retained variant",
-    "impact": "LOW",
-    "order": 15,
-  },
-  "SO_0001567": {
-    "id": "SO_0001567",
-    "key": "stop_retained_variant",
-    "label": "Stop retained variant",
-    "impact": "LOW",
-    "order": 16,
-  },
-  "SO_0001819": {
-    "id": "SO_0001819",
-    "key": "synonymous_variant",
-    "label": "Synonymous variant",
-    "impact": "LOW",
-    "order": 17,
-  },
-  "SO_0001580": {
-    "id": "SO_0001580",
-    "key": "coding_sequence_variant",
-    "label": "Coding sequence variant",
-    "impact": "MODIFIER",
-    "order": 18,
-  },
-  "SO_0001620": {
-    "id": "SO_0001620",
-    "key": "mature_miRNA_variant",
-    "label": "Mature miRNA variant",
-    "impact": "MODIFIER",
-    "order": 19,
-  },
-  "SO_0001623": {
-    "id": "SO_0001623",
-    "key": "5_prime_UTR_variant",
-    "label": "5 prime UTR variant",
-    "impact": "MODIFIER",
-    "order": 20,
-  },
-  "SO_0001624": {
-    "id": "SO_0001624",
-    "key": "3_prime_UTR_variant",
-    "label": "3 prime UTR variant",
-    "impact": "MODIFIER",
-    "order": 21,
-  },
-  "SO_0001792": {
-    "id": "SO_0001792",
-    "key": "non_coding_transcript_exon_variant",
-    "label": "Non coding transcript exon variant",
-    "impact": "MODIFIER",
-    "order": 22,
-  },
-  "SO_0001627": {
-    "id": "SO_0001627",
-    "key": "intron_variant",
-    "label": "Intron variant",
-    "impact": "MODIFIER",
-    "order": 23,
-  },
-  "SO_0001621": {
-    "id": "SO_0001621",
-    "key": "NMD_transcript_variant",
-    "label": "NMD transcript variant",
-    "impact": "MODIFIER",
-    "order": 24,
-  },
-  "SO_0001619": {
-    "id": "SO_0001619",
-    "key": "non_coding_transcript_variant",
-    "label": "Non coding transcript variant",
-    "impact": "MODIFIER",
-    "order": 25,
-  },
-  "SO_0001631": {
-    "id": "SO_0001631",
-    "key": "upstream_gene_variant",
-    "label": "Upstream gene variant",
-    "impact": "MODIFIER",
-    "order": 26,
-  },
-  "SO_0001632": {
-    "id": "SO_0001632",
-    "key": "downstream_gene_variant",
-    "label": "Downstream gene variant",
-    "impact": "MODIFIER",
-    "order": 27,
-  },
-  "SO_0001895": {
-    "id": "SO_0001895",
-    "key": "TFBS_ablation",
-    "label": "TFBS ablation",
-    "impact": "MODIFIER",
-    "order": 28,
-  },
-  "SO_0001892": {
-    "id": "SO_0001892",
-    "key": "TFBS_amplification",
-    "label": "TFBS amplification",
-    "impact": "MODIFIER",
-    "order": 29,
-  },
-  "SO_0001782": {
-    "id": "SO_0001782",
-    "key": "TF_binding_site_variant",
-    "label": "TF binding site variant",
-    "impact": "MODIFIER",
-    "order": 30,
-  },
-  "SO_0001894": {
-    "id": "SO_0001894",
-    "key": "regulatory_region_ablation",
-    "label": "Regulatory region ablation",
-    "impact": "MODERATE",
-    "order": 31,
-  },
-  "SO_0001891": {
-    "id": "SO_0001891",
-    "key": "regulatory_region_amplification",
-    "label": "Regulatory region amplification",
-    "impact": "MODIFIER",
-    "order": 32,
-  },
-  "SO_0001907": {
-    "id": "SO_0001907",
-    "key": "feature_elongation",
-    "label": "Feature elongation",
-    "impact": "MODIFIER",
-    "order": 33,
-  },
-  "SO_0001566": {
-    "id": "SO_0001566",
-    "key": "regulatory_region_variant",
-    "label": "Regulatory region variant",
-    "impact": "MODIFIER",
-    "order": 34,
-  },
-  "SO_0001906": {
-    "id": "SO_0001906",
-    "key": "feature_truncation",
-    "label": "Feature truncation",
-    "impact": "MODIFIER",
-    "order": 35,
-  },
-  "SO_0001628": {
-    "id": "SO_0001628",
-    "key": "intergenic_variant",
-    "label": "Intergenic variant",
-    "impact": "MODIFIER",
-    "order": 36,
-  },
-};
-
-export const CLINICAL_SIGNIFICANCE = {
-  "pathogenic": {
-    "key": "P",
-  },
-  "likely pathogenic": {
-    "key": "LP",
-  },
-  "uncertain significance": {
-    "key": "US",
-  },
-  "likely benign": {
-    "key": "LB",
-  },
-  "benign": {
-    "key": "B",
-  },
-  "conflicting interpretations of pathogenicity": {
-    "key": "CI",
-  },
-  "drug response": {
-    "key": "DR",
-  },
-  "association": {
-    "key": "A",
-  },
-  "risk factor": {
-    "key": "RF",
-  },
-  "protective": {
-    "key": "PR",
-  },
-  "affects": {
-    "key": "AF",
-  },
-  "other": {
-    "key": "O",
-  },
-  "not provided": {
-    "key": "NP",
-  },
-  "association_not found": {
-    "key": "AN",
-  },
-};
-
-export const REVIEW_STATUS = {
-  "no assertion provided": {
-    "stars": 0,
-  },
-  "no assertion criteria provided": {
-    "stars": 0,
-  },
-  "no assertion for the individual variant": {
-    "stars": 0,
-  },
-  "criteria provided, single submitter": {
-    "stars": 1,
-  },
-  "criteria provided, conflicting interpretations": {
-    "stars": 1,
-  },
-  "criteria provided, multiple submitters, no conflicts": {
-    "stars": 2,
-  },
-  "reviewed by expert panel": {
-    "stars": 3,
-  },
-  "practice guideline": {
-    "stars": 4,
-  },
-};
-
-export const ALPHAMISSENSE = {
-  "likely_pathogenic": {
-    "key": "LP",
-    "label": "Likely pathogenic"
-  },
-  "ambiguous": {
-    "key": "AMBIGUOUS",
-    "label": "Ambiguous"
-  },
-  "likely_benign": {
-    "key": "LB",
-    "label": "Likely benign"
-  }
-};
-
-export const SIFT = {
-  "deleterious": {
-    "key": "D",
-    "label": "Deleterious"
-  },
-  "tolerated": {
-    "key": "T",
-    "label": "Tolerated"
-  },
-};
-
-export const POLYPHEN = {
-  "probably_damaging": {
-    "key": "PROBD",
-    "label": "Probably Damaging"
-  },
-  "possibly_damaging": {
-    "key": "POSSD",
-    "label": "Possibly Damaging"
-  },
-  "benign": {
-    "key": "B",
-    "label": "Benign"
-  },
-  "unknown": {
-    "key": "U",
-    "label": "Unknown"
-  },
-};
